--- async-race/src/modules/components/textInput.ts
@@ -1,14 +1,0 @@
-import { CarData } from '../utils';
-
-export const createTextInput = (
-  className: string,
-  state: CarData
-): HTMLInputElement => {
-  const input = document.createElement('input');
-  input.classList.add('input-text');
-  input.classList.add(className);
-
-  input.addEventListener('change', (): string => (state.name = input.value));
-
-  return input;
-};
